Fix LimeSDR Mini v2 radio start
This fixes the issue when starting LimeSDR Mini v2 radio does not
really start the actual reception. The workaround was to configure
the bandwidth after the radio was started.

The actual fix is to enable TX channel. This is how the demo code
from LimeSuite is communicating with the v2 of the radio, and this
is also was confirmed on the myriad discourse:

  https://discourse.myriadrf.org/t/limesdr-mini-v2-reading-samples-from-rx-fifo/7900/2

The TX channel is only enabled for the LimeSDR Mini v2 and not for
other LimeSDR radios to avoid unwanted possible side effects. The
check is performed based on the name, matching the LimeSDRTest
application from the LimeSuite. The name of the board is inlined
to avoid dependency of the new LMSBoard.h.

Upon the radio stop both RXC and TX channels are stopped. This is
for the code simplicity. Alternative would be to rely in the
LMS_Close() to disabled channels which matches the basicRX demo
code from the LimeSuite. It is not really clear what is the best
strategy here.