Make it easier to contribute: add .editorconfig
http://editorconfig.org/
The .editorconfig, which is a common way to tell IDEs and editors
basic expectations such as indentation of code. Almost every
editor supports that and people have it configured.

Set the configuration to the observed 4 space indent for c++ code.

That way, code contributions will automatically adhere to the
basic formatting even if not evaluating .clang-format.